+++ core/src/services/navigation.js
@@ -1,5 +1,19 @@
 import { buildNode } from './navigation_helpers';
+import { getConfigValue, getConfigValueFromObject } from './config';
+
+import {
+  replaceVars,
+  hasIframeIsolation,
+  hideElementChildren,
+  removeElementChildren,
+  isNotSameDomain
+} from '../utilities/helpers';
 import * as CHILDREN from './navigation_children';
+
+let timeoutHandle;
+const contextVarPrefix = 'context.';
+const nodeParamsVarPrefix = 'nodeParams.';
+const iframeNavFallbackTimeout = 2000;
 
 export const getNavigationPath = async (rootNavProviderPromise, activePath) => {
   const rootNode = {};
@@ -130,4 +144,130 @@ export const getLeftNavData = async (current, componentData) => {
     updatedCompData.children = groupedChildren;
   }
   return updatedCompData;
+};
+
+/**
+  navigateTo used for navigation
+  @param route string  absolute path of the new route
+  @param options object  navi options, eg preserveView
+  @param windowElem object  defaults to window
+  @param documentElem object  defaults to document
+ */
+export const navigateTo = (route, windowElem = window) => {
+  if (getConfigValue('routing.useHashRouting')) {
+    windowElem.location.hash = route;
+    return;
+  }
+
+  windowElem.history.pushState(
+    {
+      path: route
+    },
+    '',
+    route
+  );
+
+  // https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent#Browser_compatibility
+  // https://developer.mozilla.org/en-US/docs/Web/API/Event#Browser_compatibility
+  // https://developer.mozilla.org/en-US/docs/Web/API/Event/createEvent
+  let event;
+  if (isIE()) {
+    event = document.createEvent('Event');
+    event.initEvent('popstate', true, true);
+  } else {
+    event = new CustomEvent('popstate');
+  }
+
+  windowElem.dispatchEvent(event);
+};
+
+export const navigateIframe = (config, component, node) => {
+  clearTimeout(timeoutHandle);
+  const componentData = component.get();
+  let viewUrl = componentData.viewUrl;
+  if (viewUrl) {
+    viewUrl = replaceVars(viewUrl, componentData.context, contextVarPrefix);
+    viewUrl = replaceVars(
+      viewUrl,
+      componentData.nodeParams,
+      nodeParamsVarPrefix
+    );
+  }
+
+  if (
+    !componentData.isNavigateBack &&
+    (isNotSameDomain(config, component) ||
+      hasIframeIsolation(component) ||
+      Boolean(config.builderCompatibilityMode))
+  ) {
+    const componentData = component.get();
+    // preserveView, hide other frames, else remove
+    if (config.iframe === null) {
+      hideElementChildren(node);
+    } else {
+      removeElementChildren(node);
+    }
+
+    if (componentData.viewUrl) {
+      if (
+        getConfigValueFromObject(
+          componentData,
+          'currentNode.loadingIndicator.enabled'
+        ) !== false
+      ) {
+        component.set({ showLoadingIndicator: true });
+      } else {
+        component.set({ showLoadingIndicator: false });
+      }
+      config.navigateOk = undefined;
+      config.iframe = document.createElement('iframe');
+      config.iframe.src = viewUrl;
+
+      node.prepend(config.iframe);
+
+      if (config.builderCompatibilityMode) {
+        config.iframe.addEventListener('load', () => {
+          window.postMessage({ msg: 'luigi.hide-loading-indicator' }, '*');
+          config.iframe.contentWindow.postMessage(
+            ['init', JSON.stringify(componentData.context)],
+            '*'
+          );
+        });
+      }
+    }
+  } else {
+    const goBackContext = component.get().goBackContext;
+    config.iframe.style.display = 'block';
+    config.iframe.contentWindow.postMessage(
+      {
+        msg: 'luigi.navigate',
+        viewUrl: viewUrl,
+        context: JSON.stringify(
+          Object.assign({}, componentData.context, { goBackContext })
+        ),
+        nodeParams: JSON.stringify(Object.assign({}, componentData.nodeParams)),
+        pathParams: JSON.stringify(Object.assign({}, componentData.pathParams)),
+        internal: JSON.stringify(component.prepareInternalData())
+      },
+      '*'
+    );
+    // clear goBackContext after sending it to the client
+    component.set({ goBackContext: undefined });
+
+    /**
+     * check if luigi responded
+     * if not, callback again to replace the iframe
+     */
+    timeoutHandle = setTimeout(() => {
+      if (config.navigateOk) {
+        config.navigateOk = undefined;
+      } else {
+        config.iframe = undefined;
+        console.info(
+          'navigate: luigi-client did not respond, using fallback by replacing iframe'
+        );
+        navigateIframe(config, component, node);
+      }
+    }, iframeNavFallbackTimeout);
+  }
 };
