+++ core/src/services/routing.js
@@ -1,60 +1,15 @@
-import { getNavigationPath } from './navigation';
-import {
-  getConfigValue,
-  getConfigValueAsync,
-  getConfigBooleanValue,
-  getConfigValueFromObject
-} from './config';
+import { getNavigationPath, navigateTo, navigateIframe } from './navigation';
+import { getConfigValueAsync, getConfigBooleanValue } from './config';
 import {
   getPathWithoutHash,
-  getUrlWithoutHash,
-  isIE
+  parseParams,
+  getNodeParams,
+  getPathParams,
+  getLastNodeObject,
+  getDefaultChildNode,
+  isExistingRoute,
+  buildRoute
 } from '../utilities/helpers';
-
-const iframeNavFallbackTimeout = 2000;
-let timeoutHandle;
-const defaultContentViewParamPrefix = '~';
-
-const getLastNodeObject = pathData => {
-  const lastElement = [...pathData.navigationPath].pop();
-  return lastElement ? lastElement : {};
-};
-
-const getDefaultChildNode = function(pathData) {
-  const lastElement =
-    pathData.navigationPath[pathData.navigationPath.length - 1];
-
-  const pathExists = lastElement.children.find(
-    childNode => childNode.pathSegment === lastElement.defaultChildNode
-  );
-
-  if (lastElement.defaultChildNode && pathExists) {
-    return lastElement.defaultChildNode;
-  } else {
-    return lastElement.children[0].pathSegment;
-  }
-};
-
-const isExistingRoute = function(path, pathData) {
-  if (!path) {
-    return true;
-  }
-
-  const lastElement =
-    pathData.navigationPath[pathData.navigationPath.length - 1];
-  const routeSplit = path.replace(/\/$/, '').split('/');
-  const lastPathSegment = routeSplit[routeSplit.length - 1];
-
-  return lastElement.pathSegment === lastPathSegment;
-};
-
-const hideElementChildren = node => {
-  if (node.children) {
-    Array.from(node.children).forEach(child => {
-      child.style.display = 'none';
-    });
-  }
-};
 
 export const getActiveIframe = node => {
   return node.firstChild;
@@ -81,200 +36,6 @@ export const removeInactiveIframes = node => {
     }
   });
 };
-
-const removeElementChildren = node => {
-  while (node.firstChild) {
-    node.removeChild(node.firstChild);
-  }
-};
-
-export const isNotSameDomain = (config, component) => {
-  if (config.iframe) {
-    const componentData = component.get();
-    const previousUrl = getUrlWithoutHash(
-      componentData.previousNodeValues.viewUrl
-    );
-    const nextUrl = getUrlWithoutHash(componentData.viewUrl);
-    return previousUrl != nextUrl;
-  }
-  return true;
-};
-
-export const hasIframeIsolation = component => {
-  const componentData = component.get();
-  return (
-    componentData.isolateView || componentData.previousNodeValues.isolateView
-  );
-};
-
-export const getContentViewParamPrefix = () => {
-  return (
-    getConfigValue('routing.contentViewParamPrefix') ||
-    defaultContentViewParamPrefix
-  );
-};
-
-const contextVarPrefix = 'context.';
-const nodeParamsVarPrefix = 'nodeParams.';
-
-const escapeRegExp = string => {
-  return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-};
-
-const replaceVars = (viewUrl, params, prefix) => {
-  let processedUrl = viewUrl;
-  if (params) {
-    Object.entries(params).forEach(entry => {
-      processedUrl = processedUrl.replace(
-        new RegExp(escapeRegExp('{' + prefix + entry[0] + '}'), 'g'),
-        encodeURIComponent(entry[1])
-      );
-    });
-  }
-  processedUrl = processedUrl.replace(
-    new RegExp('\\{' + escapeRegExp(prefix) + '[^\\}]+\\}', 'g'),
-    ''
-  );
-  return processedUrl;
-};
-
-const navigateIframe = (config, component, node) => {
-  clearTimeout(timeoutHandle);
-  const componentData = component.get();
-  let viewUrl = componentData.viewUrl;
-  if (viewUrl) {
-    viewUrl = replaceVars(viewUrl, componentData.context, contextVarPrefix);
-    viewUrl = replaceVars(
-      viewUrl,
-      componentData.nodeParams,
-      nodeParamsVarPrefix
-    );
-  }
-
-  if (
-    !componentData.isNavigateBack &&
-    (isNotSameDomain(config, component) ||
-      hasIframeIsolation(component) ||
-      Boolean(config.builderCompatibilityMode))
-  ) {
-    const componentData = component.get();
-    // preserveView, hide other frames, else remove
-    if (config.iframe === null) {
-      hideElementChildren(node);
-    } else {
-      removeElementChildren(node);
-    }
-
-    if (componentData.viewUrl) {
-      if (
-        getConfigValueFromObject(
-          componentData,
-          'currentNode.loadingIndicator.enabled'
-        ) !== false
-      ) {
-        component.set({ showLoadingIndicator: true });
-      } else {
-        component.set({ showLoadingIndicator: false });
-      }
-      config.navigateOk = undefined;
-      config.iframe = document.createElement('iframe');
-      config.iframe.src = viewUrl;
-
-      node.prepend(config.iframe);
-
-      if (config.builderCompatibilityMode) {
-        config.iframe.addEventListener('load', () => {
-          window.postMessage({ msg: 'luigi.hide-loading-indicator' }, '*');
-          config.iframe.contentWindow.postMessage(
-            ['init', JSON.stringify(componentData.context)],
-            '*'
-          );
-        });
-      }
-    }
-  } else {
-    const goBackContext = component.get().goBackContext;
-    config.iframe.style.display = 'block';
-    config.iframe.contentWindow.postMessage(
-      {
-        msg: 'luigi.navigate',
-        viewUrl: viewUrl,
-        context: JSON.stringify(
-          Object.assign({}, componentData.context, { goBackContext })
-        ),
-        nodeParams: JSON.stringify(Object.assign({}, componentData.nodeParams)),
-        pathParams: JSON.stringify(Object.assign({}, componentData.pathParams)),
-        internal: JSON.stringify(component.prepareInternalData())
-      },
-      '*'
-    );
-    // clear goBackContext after sending it to the client
-    component.set({ goBackContext: undefined });
-
-    /**
-     * check if luigi responded
-     * if not, callback again to replace the iframe
-     */
-    timeoutHandle = setTimeout(() => {
-      if (config.navigateOk) {
-        config.navigateOk = undefined;
-      } else {
-        config.iframe = undefined;
-        console.info(
-          'navigate: luigi-client did not respond, using fallback by replacing iframe'
-        );
-        navigateIframe(config, component, node);
-      }
-    }, iframeNavFallbackTimeout);
-  }
-};
-
-const parseParams = paramsString => {
-  const result = {};
-  const viewParamString = paramsString;
-  const pairs = viewParamString ? viewParamString.split('&') : null;
-  if (pairs) {
-    pairs.forEach(pairString => {
-      const keyValue = pairString.split('=');
-      if (keyValue && keyValue.length > 0) {
-        result[decodeURIComponent(keyValue[0])] = decodeURIComponent(
-          keyValue[1]
-        );
-      }
-    });
-  }
-  return result;
-};
-
-const getNodeParams = params => {
-  const result = {};
-  const paramPrefix = getContentViewParamPrefix();
-  if (params) {
-    Object.entries(params).forEach(entry => {
-      if (entry[0].startsWith(paramPrefix)) {
-        const paramName = entry[0].substr(paramPrefix.length);
-        result[paramName] = entry[1];
-      }
-    });
-  }
-  return result;
-};
-
-const getPathParams = nodes => {
-  const params = {};
-  nodes
-    .filter(n => n.pathParam)
-    .map(n => n.pathParam)
-    .forEach(pp => {
-      params[pp.key.replace(':', '')] = pp.value;
-    });
-  return params;
-};
-
-const buildRoute = (node, path, params) =>
-  !node.parent
-    ? path + (params ? '?' + params : '')
-    : buildRoute(node.parent, `/${node.parent.pathSegment}${path}`, params);
 
 export const handleRouteChange = async (path, component, node, config) => {
   try {
@@ -372,41 +133,6 @@ export const matchPath = async path => {
     console.error('Could not match path', err);
   }
   return null;
-};
-
-/**
-  navigateTo used for navigation
-  @param route string  absolute path of the new route
-  @param options object  navi options, eg preserveView
-  @param windowElem object  defaults to window
-  @param documentElem object  defaults to document
- */
-export const navigateTo = (route, windowElem = window) => {
-  if (getConfigValue('routing.useHashRouting')) {
-    windowElem.location.hash = route;
-    return;
-  }
-
-  windowElem.history.pushState(
-    {
-      path: route
-    },
-    '',
-    route
-  );
-
-  // https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent#Browser_compatibility
-  // https://developer.mozilla.org/en-US/docs/Web/API/Event#Browser_compatibility
-  // https://developer.mozilla.org/en-US/docs/Web/API/Event/createEvent
-  let event;
-  if (isIE()) {
-    event = document.createEvent('Event');
-    event.initEvent('popstate', true, true);
-  } else {
-    event = new CustomEvent('popstate');
-  }
-
-  windowElem.dispatchEvent(event);
 };
 
 export const handleRouteClick = (node, windowElem = window) => {
